service/requirements.txt
aniso8601
attrs
Click
Flask==1.1.2
Flask-Cors
flask-restplus
itsdangerous
Jinja2
joblib
jsonschema
MarkupSafe
mysql-connector-python
numpy
pandas
pyrsistent
pytz
scikit-learn
scipy
six
sklearn
Werkzeug==1.0.1
